refactor: 523 - move structure from OptionItems to component
