--- src/widgets/option-item/index.ts
@@ -1,1 +1,0 @@
-export { OptionItem } from './ui/option-item';
