--- src/widgets/option-item/ui/option-item.module.scss
@@ -1,14 +1,0 @@
-.option-item {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  max-width: 500px;
-
-  @include media-laptop {
-    max-width: 600px;
-  }
-
-  @include media-tablet {
-    max-width: 100%;
-  }
-}
